requirements.txt
google-generativeai==0.8.0
numpy==1.26.4
pandas==2.2.2
requests==2.32.3
PyPDF2==3.0.1
Pillow==10.3.0
pdf2image==1.17.0
pdfplumber==0.11.0
pytesseract==0.3.10
requests==2.32.3
